# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/agents/context": {
      "get": {
        "operationId": "IntelligenceController_context",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "预测预警与智能助手"
        ]
      }
    },
...
    "/api/billing/bills": {
      "get": {
        "operationId": "BillingController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "对账结算"
        ]
      },
      "post": {
        "operationId": "BillingController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BillDraft"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "对账结算"
        ]
      }
    },
...
    "/api/contracts": {
      "get": {
        "operationId": "ContractsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "合同包与模拟签署"
        ]
      },
      "post": {
        "operationId": "ContractsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateContract"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "合同包与模拟签署"
        ]
      }
    },
...
    "/api/fulfillment": {
      "get": {
        "operationId": "FulfillmentController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "联运执行"
        ]
      },
      "post": {
        "operationId": "FulfillmentController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BusinessDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "联运执行"
        ]
      }
    },
...
    "/api/fulfillment/issues/{id}/resolve": {
      "post": {
        "operationId": "FulfillmentController_resolve",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolutionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "联运执行"
        ]
      }
    },
...
    "/api/trade-orders/import": {
      "post": {
        "operationId": "TransportController_import",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportOrderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "平台幂等导入交易订单，不覆盖已有来源事实",
        "tags": [
          "交易订单与供需管理"
        ]
      }
    },
...
    "/api/plans/runs/{id}": {
      "get": {
        "operationId": "PlanningController_run",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "联运求解与线路价格"
        ]
      }
    },
...
    "/api/map/config": {
      "get": {
        "operationId": "MapController_config",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "地图与地点"
        ]
      }
    },
...
    "/api/_AMapService/{path}": {
      "get": {
        "operationId": "MapController_proxy",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "地图与地点"
        ]
      }
    },
...
  },
  "info": {
    "title": "北粮南运 · 智能多式联运 API",
    "description": "Phase 9：需求、方案、匹配、模拟签章、联运执行、预测预警、智能辅助、账单版本、双方结算确认和授权数据服务。",
    "version": "0.9.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "BillDraft": {
        "type": "object",
        "properties": {}
      },
...
      "CreateContract": {
        "type": "object",
        "properties": {}
      },
...
      "BusinessDto": {
        "type": "object",
        "properties": {}
      },
...
      "ResolutionDto": {
        "type": "object",
        "properties": {}
      },
...
      "OrderItemDto": {
        "type": "object",
        "properties": {
          "lineNo": {
            "type": "string"
          },
          "grainId": {
            "type": "string"
          },
          "cargoName": {
            "type": "string"
          },
          "specification": {
            "type": "string"
          },
          "quantity": {
            "type": "number"
          }
        },
        "required": [
          "lineNo",
          "grainId",
          "cargoName",
          "specification",
          "quantity"
        ]
      },
      "ImportOrderDto": {
        "type": "object",
        "properties": {
          "businessNo": {
            "type": "string"
          },
          "businessEntityId": {
            "type": "string"
          },
          "recipient": {
            "type": "string"
          },
          "shipperContact": {
            "type": "string"
          },
          "shipperPhone": {
            "type": "string"
          },
          "recipientContact": {
            "type": "string"
          },
          "recipientPhone": {
            "type": "string"
          },
          "sourceSystem": {
            "type": "string"
          },
          "sourceRecordId": {
            "type": "string"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItemDto"
            }
          }
        },
        "required": [
          "businessNo",
          "businessEntityId",
          "recipient",
          "shipperContact",
          "shipperPhone",
          "recipientContact",
          "recipientPhone",
          "sourceSystem",
          "sourceRecordId",
          "items"
        ]
      },
...
    }
  }
}